pyproject.toml
[tool.poetry]
name = "trade_flow"
version = "0.1.0"
description = "TradeFlow is an intelligent trading bot that leverages automation and analysis to navigate markets for you"
authors = ["Fortesense Labs <fortesenselabs@gmail.com>"]
readme = "README.md"
keywords = [
    "algo",
    "trading",
    "deep learning",
    "machine learning",
    "ai",
    "Reinforcement Learning",
    "Finance",
]
classifiers = [
    "Programming Language :: Python",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.7",
    "Programming Language :: Python :: 3.8",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: Implementation :: CPython",
    "Programming Language :: Python :: Implementation :: PyPy",
    "Topic :: Software Development :: Libraries",
    "Topic :: Software Development :: Libraries :: Python Modules",
    "Topic :: Scientific/Engineering",
    "Topic :: Office/Business :: Financial",
    "Topic :: Office/Business :: Financial :: Investment",
    "Intended Audience :: Developers",
    "Intended Audience :: Financial and Insurance Industry",
    "Intended Audience :: Science/Research",
]

# include = [
#     # Rust source must be included in the source distributions
#     { path = "nautilus_core/*", format = "sdist" },
#     { path = "nautilus_core/*[!target]*/*", format = "sdist" },
#     # Compiled extensions must be included in the wheel distributions
#     { path = "nautilus_trader/**/*.so", format = "wheel" },
#     { path = "nautilus_trader/**/*.pyd", format = "wheel" },
#     # Include the py.typed file for type checking support
#     { path = "nautilus_trader/py.typed", format = "sdist" },
#     { path = "nautilus_trader/py.typed", format = "wheel" },
#     # Include Python interface files for type checking support
#     { path = "nautilus_trader/**/*.pyi", format = "sdist" },
#     { path = "nautilus_trader/**/*.pyi", format = "wheel" },
# ]


[build-system]
requires = [
    "setuptools>=64",
    "poetry-core>=1.9.0",
    # "numpy>=2.0.1",
    # "Cython==3.0.11",
    "toml>=0.10.2",
]
build-backend = "poetry.core.masonry.api"

[tool.setuptools]
packages = [
    "trade_flow",
    "packages.nautilus_mt5api",
    "packages.adapters.metatrader5",
]

[tool.poetry.scripts]
flow = "trade_flow.daemon.server:run"
flowcli = "trade_flow.cli.main:cli"

# [tool.poetry.build]
# script = "build.py"
# generate-setup-file = false

[tool.poetry.dependencies]
python = ">=3.11,<3.13"
ta-lib = "^0.4.32"                                                                           # Python wrapper for TA-lib (native) library
elegantrl = { git = "https://github.com/AI4Finance-Foundation/ElegantRL.git#egg=elegantrl" }
tsfresh = "^0.20.3"                                                                          # Features/label generation | if tsfresh feature generator is used. It depends on matrixprofile with binary (in pip) only for Python 3.8
scikit-learn = "^1.5.1"
lightgbm = "^4.5.0"
stable-baselines3 = { extras = ["extra"], version = "^2.3.2" }
ray = { extras = ["default", "rllib", "serve", "tune"], version = "^2.35.0" }
torch = "^2.4.0"
seaborn = "^0.13.2"
colorlog = "^6.8.2"
requests = "^2.32.3"
websockets = "^12.0"
pydantic = "^2.8.2"
python-dotenv = "^1.0.1"
pyyaml = "^6.0.2"
cachetools = "4.2.2"
sqlalchemy = "2.0.31"
flask = "3.0.0"
flask-jsonrpc = "1.1.0"
jsonrpcclient = "^4.0.3"
jsonrpcserver = "^5.0.9"
jsonschema = "4.20.0"
jsonschema-specifications = "2023.12.1"
apscheduler = "3.10"
deprecated = ">=1.2.13"
matplotlib = ">=3.1.1"
plotly = ">=4.5.0"
click = "^8.1.7"
docker = "^7.1.0"
rpyc = "^6.0.0"
tqdm = "^4.66.5"
typing-extensions = "4.9.0"
typeguard = "2.13.3"
tabulate = "0.9.0"
ipython = "^8.27.0"
toml = "^0.10.2"
pandas-ta = "^0.3.14b0"
pandas = "^2.2.2"
feature-engine = "^1.8.1"
quantstats = "^0.0.62"
optuna = "^4.0.0"
tensorboard = "^2.17.1"
# pandas = "^2.2.2"                                                            >
# numba = "^0.60.0"   # For highlow label generator. numba 0.57.1 requires numpy<>
# pyarrow = "^17.0.0" # If parquet is used. Alternatively fastparquet
# ta-lib = "^0.4.32"  # Python wrapper for TA-lib (native) library
# pandas-ta = "^0.3.14b0"
# ta = "^0.11.0"
# smartmoneyconcepts = "^0.0.24"
# nautilus-trader = "^1.198.0"
# tensorflow = "2.14.1"
# polars = "^1.6.0"
# polars-talib = "^0.1.3"
# numpy = "1.26.4"      
# numba = "^0.59.1"
nautilus-trader = { extras = ["ib"], version = "^1.201.0" }
scipy = "^1.14.1"
invoke = "^2.2.0"

[tool.poetry.extras]
docker = ["docker"]
# mt5 = ["nautilus_mt5api", "async-timeout", "defusedxml"]

[tool.poetry.group.dev]
optional = true

[tool.poetry.group.dev.dependencies]
black = "^24.8.0"
docformatter = "^1.7.5"
mypy = "^1.11.1"
pandas-stubs = "^2.2.2"
pre-commit = "^3.8.0"
ruff = "^0.5.7"
types-pytz = "^2024.1"
types-requests = "^2.32"
types-toml = "^0.10.2"
hvplot = "^0.10.0"
yfinance = {extras = ["optional"], version = "^0.2.43"}

[tool.poetry.group.mt5]
optional = true

[tool.poetry.group.mt5.dependencies]
MetaTrader5 = "^5.0"

[tool.poetry.group.test]
optional = true

[tool.poetry.group.test.dependencies]
coverage = "^7.6.1"
pytest = "^8.3.2"
requests-mock = "^1.12.1"

[tool.poetry.group.docs]
optional = true

[tool.poetry.group.docs.dependencies]
numpydoc = "^1.7.0"
linkify-it-py = "^2.0.3"
myst-parser = "^3.0.1"
sphinx_comments = "^0.0.3"
sphinx_markdown_builder = "^0.6.6"

[tool.black]
target_version = ["py310", "py311", "py312"]
line_length = 100

[tool.docformatter]
black = true
make-summary-multi-line = true
pre-summary-new-line = true
blank = true
recursive = true
in-place = true

[tool.ruff]
extend-exclude = []
line-length = 100
indent-width = 4

[tool.ruff.lint]
select = [
    # pycodestyle
    "E",
    # Pyflakes
    "F",
    # pyupgrade
    "UP",
    # flake8-bugbear
    "B",
    # flake8-simplify
    "SIM",
    # isort
    "I",
]

[tool.pytest.ini_options]
testpaths = ["tests"]
pythonpath = [".", "trade_flow"]
asyncio_mode = "strict"
filterwarnings = ["ignore::UserWarning", "ignore::DeprecationWarning"]

ignore = ["E501"] # Line too long
